folder_update_prepper: account properly for indirect dir pointers
Manipulating a multi-level directory block tree ends up adding and
unreferencing pointers, and we need to track these in the MD in the
same way that we do for files.  So track the unreferenced pointers in
`folderBlockOps` as they are dirtied, and reference new pointers made
as part of readying the directory.

Also, `StateChecker` now properly checks indirect directory blocks.

To test this, I locally enabled directory block-splitting for a
complicated concurrency test, but I don't want to check that in quite
yet because that particular test gets very slow; I want to build a
real way to test via the DSL tests first.

Issue: KBFS-3302